Move classes to their own files; update member visibility
