+++ features/mpabout/src/main/java/com/jpp/mpabout/licenses/GoToLicenseContentEvent.kt
@@ -1,0 +1,7 @@
+package com.jpp.mpabout.licenses
+
+/**
+ * Represents the navigation event to redirect the user to the license
+ * content details.
+ */
+internal data class GoToLicenseContentEvent(val licenseId: Int)
