+++ features/mpabout/src/main/java/com/jpp/mpabout/licenses/LicenseItem.kt
@@ -1,0 +1,9 @@
+package com.jpp.mpabout.licenses
+
+/**
+ * Represents an item in the list of results shown.
+ */
+internal data class LicenseItem(
+    val id: Int,
+    val name: String
+)
